Outlined tile destruction
Created a TileDestroyer class I'll use to hold blocks to be destroyed.
So far, I'm thinking a TileDestroyer will be created for every
contiguous region. On update, if the wiper is past the rightmost tile in
the region, then all tiles in the TileDestroyer will get killed. Still
thinking about how to tell any tiles above those regions to drop.